Fix Loan Edit screen, update dependencies, and resolve test failures

+++ lib/presentation/screens/dashboard/dashboard_components/wealth_hero.dart
@@ -243,7 +243,7 @@ class _WealthHeroState extends ConsumerState<WealthHero> {
                         curve: Curves.easeOutExpo,
                         builder: (context, value, child) {
                           final text = CurrencyFormatter.format(
-                            value.toInt(),
+                            value,
                             compact: false,
                             isPrivate: isPrivate,
                           );
